moved around some depictors, updated to centralize depiction
