+++ peepingtom/visualisation/depictors/__init__.py
@@ -1,3 +1,4 @@
+from .base import Depictor
 from .images import ImageDepictor
 from .lines import LineDepictor
 from .particles import ParticleDepictor
