+++ peepingtom/visualisation/depictors/images.py
@@ -1,6 +1,6 @@
 from napari.layers import Image
 
-from ..base import Depictor
+from .base import Depictor
 
 
 class ImageDepictor(Depictor):
@@ -8,7 +8,7 @@ class ImageDepictor(Depictor):
         ikwargs = {}
         ikwargs.update(image_kwargs)
 
-        layer = Image(self.datablock.data,
-                      name=f'{self.name} - image',
-                      **ikwargs)
+        layer = self.make_image_layer(self.datablock.data,
+                                      name=f'{self.name} - image',
+                                      **ikwargs)
         self.layers.append(layer)
