+++ peepingtom/visualisation/depictors/particles.py
@@ -1,7 +1,7 @@
 import numpy as np
 from napari.layers import Points, Vectors
 
-from ..base import Depictor
+from .base import Depictor
 
 
 class ParticleDepictor(Depictor):
@@ -12,10 +12,10 @@ class ParticleDepictor(Depictor):
         pkwargs.update(point_kwargs)
         vkwargs.update(vector_kwargs)
 
-        p_layer = Points(self.datablock.positions.zyx,
-                         name=f'{self.name} - particle positions',
-                         properties=self.datablock.properties.data,
-                         **pkwargs)
+        p_layer = self.make_points_layer(self.datablock.positions.zyx,
+                                         name=f'{self.name} - particle positions',
+                                         properties=self.datablock.properties.data,
+                                         **pkwargs)
         self.layers.append(p_layer)
 
         # get positions and 'projection' vectors
@@ -23,12 +23,10 @@ class ParticleDepictor(Depictor):
         unit_z_rotated_order_xyz = self.datablock.orientations.oriented_vectors('z').reshape((-1, 3))
         unit_z_rotated_order_zyx = unit_z_rotated_order_xyz[:, ::-1]
 
-        napari_vectors = np.stack([positions,
-                                   unit_z_rotated_order_zyx],
-                                  axis=1)
-        v_layer = Vectors(napari_vectors,
-                          name=f'{self.name} - particle orientations',
-                          **vkwargs)
+        napari_vectors = np.stack([positions, unit_z_rotated_order_zyx], axis=1)
+        v_layer = self.make_vectors_layer(napari_vectors,
+                                          name=f'{self.name} - particle orientations',
+                                          **vkwargs)
         self.layers.append(v_layer)
 
     @property
