+++ peepingtom/visualisation/peepers/__init__.py
@@ -1,1 +1,2 @@
+from .base import Peeper
 from .particle_peeper import ParticlePeeper
